requirements.txt
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
backoff==2.2.1
boto3==1.28.63
botocore==1.31.63
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
html5tagger==1.3.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
mccabe==0.7.0
multidict==6.0.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-aiohttp-client==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-logging==0.47b0
opentelemetry-instrumentation-pymongo==0.47b0
opentelemetry-instrumentation-redis==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-exporter-prometheus==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.24.4
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pylint==3.0.1
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
redis==5.0.1
requests==2.31.0
s3transfer==0.7.0
sanic-routing==23.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomlkit==0.12.1
tracerite==1.1.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.18.0
websockets==12.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
prometheus_client==0.20.0
bcrypt
pydantic[email]
asyncmy
cryptography
passlib